Move tests to module folders & use `index` files to export modules
