+++ server/api/collection/index.js
@@ -1,0 +1,3 @@
+import router from './collectionRouter';
+
+export default router;
