+++ server/api/image/index.js
@@ -1,0 +1,3 @@
+import router from './imageRouter';
+
+export default router;
