+++ server/api/note/index.js
@@ -1,0 +1,3 @@
+import router from './noteRouter';
+
+export default router;
