+++ server/api/recipe/index.js
@@ -1,0 +1,3 @@
+import router from './recipeRouter';
+
+export default router;
