+++ server/api/review/index.js
@@ -1,0 +1,3 @@
+import router from './reviewRouter';
+
+export default router;
