+++ server/api/tag/index.js
@@ -1,0 +1,3 @@
+import router from './tagRouter';
+
+export default router;
