+++ server/api/user/index.js
@@ -1,0 +1,3 @@
+import router from './userRouter';
+
+export default router;
