+++ server/api/user/user.test.js
@@ -8,13 +8,13 @@ import expect from 'expect';
 import request from 'supertest';
 import faker from 'faker';
 import mongoose from 'mongoose';
-import { apiV1, setup, teardown, resetDB } from '../testHelpers/testSetup';
+import { apiV1, setup, teardown, resetDB } from '../../test/setup';
 import {
   cloudinaryPostMock,
   cloudinaryCleanup,
-} from '../testHelpers/images/imageHelpers';
+} from '../../test/helpers/images/imageHelpers';
 import app from '../../index';
-import User from '../../api/user/userModel';
+import User from './userModel';
 
 const { ObjectId } = mongoose.Types;
 
@@ -791,7 +791,7 @@ describe('/users', function() {
       });
 
       it('returns an array of images', function() {
-        const image = 'server/test/testHelpers/images/images/pinkPanther.jpg';
+        const image = 'server/test/helpers/images/images/pinkPanther.jpg';
         let token;
 
         return createNewUser
@@ -838,7 +838,7 @@ describe('/users', function() {
 
     describe('POST', function() {
       it('needs jwt authorization', function() {
-        const image = 'server/test/testHelpers/images/images/pinkPanther.jpg';
+        const image = 'server/test/helpers/images/images/pinkPanther.jpg';
 
         return createNewUser
           .then(async function(res) {
@@ -859,7 +859,7 @@ describe('/users', function() {
       });
 
       it('adds an image and updates the user', function() {
-        const image = 'server/test/testHelpers/images/images/pinkPanther.jpg';
+        const image = 'server/test/helpers/images/images/pinkPanther.jpg';
 
         return createNewUser
           .then(async function(res) {
@@ -891,7 +891,7 @@ describe('/users', function() {
 
       it('rejects unauthorized file types', function() {
         // .gif
-        const image = 'server/test/testHelpers/images/images/evilBaby.gif';
+        const image = 'server/test/helpers/images/images/evilBaby.gif';
 
         return createNewUser
           .then(async function(res) {
