+++ server/auth/index.js
@@ -1,0 +1,3 @@
+import router from './authRouter';
+
+export default router;
