+++ server/test/helpers/images/imageHelpers.js
