+++ server/test/helpers/images/images/pinkPanther.jpg
